Update filename and update code to pass tests for board
