+++ .gitignore
@@ -239,3 +239,5 @@ dmypy.json
 # End of https://www.gitignore.io/api/python,pycharm+all,visualstudiocode,jupyternotebooks
 
 src/andershoest_ex/pa02/scratch_pa01_snakes.py
+
+src/andershoest_ex/pa02/scratch.py
